+++ ts/features/bonus/bpd/components/paymentMethodActivationToggle/BpdPaymentMethodToggleFactory.tsx
@@ -1,11 +1,11 @@
 import * as React from "react";
 import {
   EnableableFunctionsTypeEnum,
-  getPaymentMethodHash,
   PaymentMethod
 } from "../../../../../types/pagopa";
 import { hasFunctionEnabled } from "../../../../../utils/walletv2";
 import { HPan } from "../../store/actions/paymentMethods";
+import { getPaymentMethodHash } from "../../../../../utils/paymentMethod";
 import PaymentMethodBpdToggle from "./base/PaymentMethodBpdToggle";
 
 /**
