+++ ts/features/wallet/onboarding/bancomat/saga/networking/index.ts
@@ -6,7 +6,6 @@ import { SagaCallReturnType } from "../../../../../../types/utils";
 import { PaymentManagerClient } from "../../../../../../api/pagopa";
 import { SessionManager } from "../../../../../../utils/SessionManager";
 import {
-  getPaymentMethodHash,
   isRawBancomat,
   PaymentManagerToken
 } from "../../../../../../types/pagopa";
@@ -16,6 +15,7 @@ import {
   searchUserPans
 } from "../../store/actions";
 import { convertWalletV2toWalletV1 } from "../../../../../../utils/walletv2";
+import { getPaymentMethodHash } from "../../../../../../utils/paymentMethod";
 
 // load all bancomat abi
 export function* handleLoadAbi(
