+++ ts/utils/paymentMethod.ts
@@ -1,0 +1,30 @@
+import {
+  isRawBancomat,
+  isRawBPay,
+  isRawCreditCard,
+  isRawSatispay,
+  RawCreditCardPaymentMethod,
+  RawPaymentMethod
+} from "../types/pagopa";
+import { FOUR_UNICODE_CIRCLES } from "./wallet";
+
+export const getPaymentMethodHash = (
+  pm: RawPaymentMethod
+): string | undefined => {
+  if (isRawBancomat(pm)) {
+    return pm.info.hashPan;
+  }
+  if (isRawCreditCard(pm)) {
+    return pm.info.hashPan;
+  }
+  if (isRawSatispay(pm)) {
+    return pm.info.uuid;
+  }
+  if (isRawBPay(pm)) {
+    return pm.info.uidHash;
+  }
+  return undefined;
+};
+
+export const getTitleFromCard = (creditCard: RawCreditCardPaymentMethod) =>
+  `${FOUR_UNICODE_CIRCLES} ${creditCard.info.blurredNumber}`;
